refactor: extract api level input handling
