+++ docs/refactoring-plan-2026-04-11.md
@@ -56,6 +56,8 @@ Completed structural checkpoints:
 - `DeviceDetails` extracted from `app::state`
 - `ApiLevelCache` extracted from `app::state`
 - `src/app/state.rs` converted into `src/app/state/mod.rs` plus sibling modules
+- app helper modules extracted so far:
+  - `api_levels.rs`
 - `src/managers/android.rs` converted into `src/managers/android/mod.rs`
 - `src/managers/ios.rs` converted into `src/managers/ios/mod.rs`
 - Android helper modules extracted so far:
