+++ src/app/mod.rs
@@ -15,6 +15,7 @@ pub mod state;
 /// Event processing optimizations for improved key input handling.
 pub mod event_processing;
 
+mod api_levels;
 mod background;
 mod create_device;
 mod details;
